+++ distribution/normal/doc.go
@@ -1,0 +1,11 @@
+// Package uniform implements a uniform continuous random variable
+//
+// Uniform distribution is a probability distribution parameterized by
+// two parameters, mean (also called location) and standard deviation
+// (also called scale)
+//
+// References:
+//
+// https://en.wikipedia.org/wiki/Normal_distribution
+// http://mathworld.wolfram.com/NormalDistribution.html
+package normal